Save IFP
Hi, 

I am writing you because we think about take over this project. We are many who don't want to let this coin die. Would you be ok to let us
the keys of this project if you don't care about it anymore? Then we could try to save it.. My nickname on Discord is William618

I hope that you will accept.

Have a nice day 

William